Update infrastructure and remove un-needed code: - Change files from js to ts - Update node version to meet vercel requirements - Remove react-awesome-reveal - Remove un-used code - Update README
